sensor_portal/requirements.txt

Django==4.2
django-debug-toolbar==3.7.0
#django-dbbackup==4.0.2
gunicorn==21.2.0


# auth
bridgekeeper==0.9
pyjwt==2.9.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0

# REST
djangorestframework==3.15
djangorestframework-gis==1.0
django-filter==22.1
drf-spectacular==0.28.0

# Django extra fields
django-sizefield==2.1.0
django-timezone-field==7.1
django-encrypted-model-fields==0.6.5
drf-recaptcha==4.0.2
django-rest-passwordreset==1.5.0
django-colorfield==0.13.0
django-icon-picker==1.0.2

# celery
celery==5.5.3
redis==4.3.4
hiredis==2.2.3
django-celery-beat==2.5.0
django-celery-results==2.5.1

# SSH access
paramiko==2.11.0
scp==0.14.4

# Aerocam
aerocam_handler @ git+https://github.com/jevansbio/aerocam_handler

#Audio
soundfile==0.13.1

# Other packages
pytz==2022.1
numpy==1.26.4
pandas==2.2.3
thefuzz==0.20.0
requests==2.32.3

# testing
pytest==8.3.3
pytest-django==4.9.0
factory-boy==3.3.1

psycopg2==2.9.3
Pillow==11.2.1

# archiving
bagit==1.8.1

#docs
mkdocs==1.6.1


